+++ GitUI/CommandsDialogs/BrowseDialog/ReleaseType.cs
@@ -1,0 +1,8 @@
+﻿namespace GitUI.CommandsDialogs.BrowseDialog;
+
+internal enum ReleaseType
+{
+    Major,
+    HotFix,
+    ReleaseCandidate
+}
